--- samples/swift/ios/README.md
@@ -1,53 +1,0 @@
-# Additional samples for Swift on iOS
-
-Theses samples demonstrate more advanced techniques for speech recognition with the Microsoft Cognitive Services Speech SDK.
-For an introduction to the SDK, please refer to the [quickstart articles for speech recognition](https://docs.microsoft.com/azure/cognitive-services/speech-service/get-started-speech-to-text?pivots=programming-languages-objectivec-swift) and [speech synthesis](https://docs.microsoft.com/azure/cognitive-services/speech-service/get-started-text-to-speech?pivots=programming-languages-objectivec-swift) on the SDK documentation page for step-by-step instructions.
-
-
-## Example app with various samples
-
-This [speech sample app](./speech-samples) shows various other techniques, currently only for pronunciation assessment.
-
-
-## Prerequisites
-
-- A subscription key for the Speech service. See [Try the speech service for free](https://docs.microsoft.com/azure/cognitive-services/speech-service/get-started).
-- A Mac with Xcode installed as iOS development environment. See the [Speech SDK installation quickstart](https://learn.microsoft.com/azure/ai-services/speech-service/quickstarts/setup-platform?pivots=programming-language-swift) for details on system requirements and setup.
-
-## Get the code for the samples
-
-- [Download the sample code to your development machine.](/README.md#get-the-samples)
-
-## Get the Speech SDK for iOS
-
-**By downloading the Microsoft Cognitive Services Speech SDK, you acknowledge its license, see [Speech SDK license agreement](https://aka.ms/csspeech/license).**
-
-The Cognitive Services Speech SDK for iOS is distributed as a xcframework bundle.
-It can be used in Xcode projects as a [CocoaPod](https://cocoapods.org/), or downloaded directly from [https://aka.ms/csspeech/iosbinary](https://aka.ms/csspeech/iosbinary) and linked manually.
-If you are going to use keyword recognition in your application, note that new keyword models created on Speech Studio will require using the SDK package from [https://aka.ms/csspeech/iosbinaryembedded](https://aka.ms/csspeech/iosbinaryembedded).
-
-## Install the SDK as a CocoaPod
-
-1. Install the CocoaPod dependency manager as described in its [installation instructions](https://guides.cocoapods.org/using/getting-started.html).
-1. Navigate to the directory of the downloaded sample app (e.g. `speech-samples`) in a terminal.
-1. Run the command `pod install`. This will generate a Xcode workspace containing both the sample app and the Speech SDK as a dependency. This workspace will be used in the following.
-
-## Build the samples
-
-Open either Xcode project with the sample code.
-This loads the project in Xcode.
-Next, add your subscription details to the `<sample name>/ViewController.swift` file:
-
-1. Replace the string `YourSubscriptionKey` with your subscription key.
-2. Replace the string `YourServiceRegion` with the [region](https://docs.microsoft.com/azure/cognitive-services/speech-service/regions) associated with your subscription (for example, `westus` for the free trial subscription).
-
-To build the sample app and check if all the paths are set correctly, choose **Product** > **Build** from the menu.
-
-## Run the samples
-
-To run the sample, click the `Play` button, or select **Product** > **Run** from the menu.
-In the simulator window that opens, you can interact with the sample application.
-
-## References
-
-* [Speech SDK API reference for Objective-C](https://aka.ms/csspeech/objectivecref)
