+++ scenarios/swift/ios/language-learning/Podfile
@@ -1,0 +1,14 @@
+target 'speech-samples' do
+    platform :ios, '12.0'
+    pod 'MicrosoftCognitiveServicesSpeech-iOS', '~> 1.25'
+    pod 'Differ'
+    use_frameworks!
+    
+    post_install do |installer|
+      installer.pods_project.targets.each do |target|
+        target.build_configurations.each do |config|
+          config.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = '12.0'
+        end
+      end
+    end
+end
